+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Changed
 =======
 
 * The Learning Paths API includes start and end dates for its steps.
+* Return enrollment date in the API instead of a boolean.
 
 0.3.3 - 2025-05-23
 ******************